Move GroupFight and CupFight related entities to separate modules
