+++ ippon/authorization_views.py
@@ -4,7 +4,7 @@ from rest_framework.response import Response
 
 import ippon.models
 import ippon.models.fight
-from ippon.models import Group
+import ippon.models.group as gm
 import ippon.models.team_fight as tfm
 import ippon.models.team as tem
 import ippon.models.tournament as tm
@@ -86,7 +86,7 @@ def team_authorization(request, pk, format=None):
 
 @api_view(['GET'])
 def group_authorization(request, pk, format=None):
-    group = get_object_or_404(Group.objects.all(), pk=pk)
+    group = get_object_or_404(gm.Group.objects.all(), pk=pk)
     return has_tournament_authorization([True, False], group.group_phase.tournament.id, request)
