+++ ippon/cup_fight/cup_fight_views_test.py
