+++ ippon/group_fight/group_fight_views_test.py
