+++ ippon/group_fight/serializers.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+
+import ippon.models
+
+
+class GroupFightSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ippon.models.group_fight.GroupFight
+        fields = (
+            'id',
+            'team_fight',
+            'group'
+        )
